requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
bcrypt==4.1.2
cryptography>=3.4.8
PyJWT==2.8.0
ldap3>=2.9.1
python-ldap>=3.4.0
python-dotenv>=1.0.0